backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
pydantic==1.10.7
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.0.0
authlib==1.2.0
httpx==0.24.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6
alembic==1.10.4
starlette==0.26.1
itsdangerous==2.1.2 